Avoid converting to u128 first, go directly to f64

+++ src/decimal.rs
@@ -2663,7 +2663,7 @@ impl ToPrimitive for Decimal {
             let frac_part = mantissa % precision;
             let frac_f64 = (frac_part as f64) / (precision as f64);
             let value = sign * ((integral_part as f64) + frac_f64);
-            let round_to = (10u128.pow(self.scale())) as f64;
+            let round_to = 10f64.powf(self.scale() as f64);
             Some(value * round_to / round_to)
         }
     }
